IMPALA-9430: always pass through kerberos configs
The behaviour of kerberos-related command line flags is changed so that
their values are always passed through to underlying libraries,
even if Kerberos isn't enabled for internal communication in Impala.

This is good because:
* Various libraries that communicate with external systems may use
  kerberos for outgoing connections, if *incoming* connections are
  not authenticated.
  e.g. it might just be enabled for HMS. Having them pick up different
  kerberos settings for outgoing connections if kerberos is disabled
  for incoming connections is a little weird. This
  is a safer default that reduces chances of inadvertant
  misconfigurations.
* It matches the documentation of the flags.

Some validations are still disabled when --principal is not set,
e.g. we don't check the replay cache directory. This is to avoid
any potential regressions or startup failures on non-kerberised
clusters.

Testing:
Added unit tests for flag validation and env var setting on the
code paths that I touched.

Change-Id: If4bb311c7ab7173232aab36c5ed801f93f38f5b9
Reviewed-on: http://gerrit.cloudera.org:8080/15340
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>