IMPALA-9447: Fix for TupleRowCompareTest.DecimalTest crash
TupleRowCompareTest.DecimalTest crashes on release build
because of a faulty test. The problem occures with Decimal16Value
when the test tries to copy the values into the tuple in FillMem().

The solution is to use memcpy to avoid gcc generating unaligned
instructions like movaps for int128_t. They would raise SegmentFault
when addresses are not aligned to 16 bytes. This works with the
other types as well.

Tests:
 - Run TupleRowCompareTest.

Change-Id: If2a53c88b57cec3a0bfcf02c492fbb560a80682d
Reviewed-on: http://gerrit.cloudera.org:8080/15347
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>