IMPALA-9350: Produce Ranger audits for column masking
Support for column mask transformation was provided in IMPALA-9009.
However, the corresponding audit logs were not generated when a query
involved policies of column masking. This patch fixes the issue by
providing the Ranger plug-in with a non-null instance of
RangerBufferAuditHandler when calling evalDataMaskPolicies() so that
the corresponding audit log could be stored in the audit handler after
the evaluation of a policy involving column masking.

Testing:
- Added a FE test to verify the audit logs corresponding to column
  masking are produced.
- Verified that the corresponding audit logs are indeed produced in a
  3-node cluster with Ranger and Solr installed.
- Verified that the patch passed the exhaustive tests in the DEBUG build.

Change-Id: I9d8a1181234dcef580f68f56c24ad7e962cfe58e
Reviewed-on: http://gerrit.cloudera.org:8080/15412
Reviewed-by: Quanlong Huang <huangquanlong@gmail.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>