IMPALA-9546: Update ranger-admin-site.xml.template after RANGER-2688
This patch bumps up CDP_BUILD_NUMBER to 2244454 which contains a change
introduced by RANGER-2688. Due to this change, we added to
ranger-admin-site.xml.template a cookie-related configuration so that
the Ranger server could be properly started.

Testing:
Verified that the data loading passes and that all the Ranger-related FE
and E2E tests are successful
- when $USE_CDP_HIVE is false, and
- when $USE_CDP_HIVE is true.

Change-Id: I7750f73834368c7109965e78b147238fc6316f49
Reviewed-on: http://gerrit.cloudera.org:8080/15533
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>