IMPALA-9513: Fix TestKuduOperations.test_column_storage_attributes
When introducing Kudu date support, test_column_storage_attributes
was modified to add the date datatype. The test expects the date
to be represented in python as datetime.date. Instead, the date
is a string in python, so the test consistently fails.

This changes the test so that it expects a python string, and the
test now passes.

Change-Id: Ic198b72041fbe8fe7376c45356e484b304c6f16c
Reviewed-on: http://gerrit.cloudera.org:8080/15567
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>