IMPALA-9566: Sentry service should not be started after IMPALA-8870
Due to the incompatibility between different versions of the Guava
libraries, after bumping up Guava in IMPALA-8870, our build script is
not supposed to start up the Sentry service when starting the
minicluster because Sentry has not had its Guava bumped up yet. However,
the patch for IMPALA-8870 did not take this into consideration when
$TARGET_FILESYSTEM is s3 and thus run-all.sh still attempts to start
up Sentry in this case. This patch fixes the bug.

Testing:
- Verified that this patch passes the core tests in the DEBUG build when
  $TARGET_FILESYSTEM is s3.

Change-Id: If81846f4251fb2aa752ba8c33615cae0ab513a62
Reviewed-on: http://gerrit.cloudera.org:8080/15590
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Csaba Ringhofer <csringhofer@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>