IMPALA-7138: detect devicemapper volumes correctly
This is a more general fix for device detection that
disables the digit-stripping heuristic if a device
with the same name as a partition already exists.

E.g. for device-mapper devices, /proc/partitions might
include a partition called "dm-0" that maps directly
to the block device "dm-0" (i.e. /sys/block/dm-0).
Before this fix, we looked for a device called "dm-",
which did not exist.

Testing:
Created a dummy device-mapper device with the following command:

  sudo dmsetup create 1gb-zero --table '0 1953125 zero'

Then confirmed that the device name in the logs was correct

    dm-0 (rotational=true)

Other devices - sda, sda1, sda2, sdb, sdb1, etc on my system
were still detected the same as before.

Change-Id: I9d231bcf9105db8d4ab03586eab74e0644337a6f
Reviewed-on: http://gerrit.cloudera.org:8080/15631
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>