IMPALA-8857: Fix flaky Kudu tests with external inserts
There are several Kudu related tests that are inherently flaky because
they insert rows into a Kudu table from the python Kudu client and
then scan the table from an Impala client even though no guarantees
are made about the consistency of operations from different clients.

This patch fixes this by adding a query option
'kudu_snapshot_read_timestamp_micros' which sets the timestamp at
which to perform Kudu snapshot reads.

The tests are then modified to set the timestamp to be equal to the
latest observed timestamp returned by the client that did the inserts
(plus 1 microsecond which the Kudu python client adds automatically
in latest_observed_timestamp() for this use case).

Change-Id: I5b787f6542dc31dcd846f19576a060a89aec891d
Reviewed-on: http://gerrit.cloudera.org:8080/15633
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>